Feature/wallets screen

+++ WalletOK/src/main/java/com/theost/walletok/data/dto/WalletDto.kt
@@ -10,10 +10,10 @@ data class WalletDto(
     @SerialName("name") val name: String,
     @SerialName("currency") val currency: Currency,
     @SerialName("id") val id: Int,
-    @SerialName("money") val money: Int,
-    @SerialName("gain") val gain: Int,
-    @SerialName("lose") val lose: Int,
-    @SerialName("lose_limit") val loseLimit: Int
+    @SerialName("money") val money: Long,
+    @SerialName("gain") val gain: Long,
+    @SerialName("lose") val lose: Long,
+    @SerialName("lose_limit") val loseLimit: Long
 )
 
 fun WalletDto.mapToWallet(): Wallet {
